Factor audio extraction from video procedure to util class

+++ app/src/main/java/com/ichi2/apisample/helper/AnkiDroidHelper.java
@@ -8,24 +8,17 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.database.Cursor;
-import android.media.MediaCodec;
-import android.media.MediaExtractor;
-import android.media.MediaFormat;
-import android.media.MediaMuxer;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Environment;
 
 import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
-import androidx.core.content.FileProvider;
 
 import com.ichi2.anki.FlashCardsContract;
 import com.ichi2.anki.api.AddContentApi;
 
 import java.io.File;
-import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -45,6 +38,7 @@ public class AnkiDroidHelper {
     public static final String HIERARCHICAL_TAG_SEPARATOR = "::";
 
     public static final String DIR_MEDIA = "/collection.media/";
+    public static final String PATH_TEMP_AUDIO = Environment.getExternalStorageDirectory().getPath() + "/tempOutput.mp3";
 
     private static final String PACKAGE_ANKI = "com.ichi2.anki";
 
@@ -336,70 +330,19 @@ public class AnkiDroidHelper {
         Uri uri = Uri.parse(uriString);
         mContext.grantUriPermission(PACKAGE_ANKI, uri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
         String type = mResolver.getType(uri);
-        File tempAudioFile = null;
         if (type.startsWith("video")) {
-            if (!type.equals("video/mp4")) {
-                mContext.revokeUriPermission(uri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
-                return null;
-            }
-            try {
-                MediaMuxer mediaMuxer = null;
-                MediaExtractor mediaExtractor = new MediaExtractor();
-                final String tempAudioPath = Environment.getExternalStorageDirectory().getPath() + "/tempOutput.mp3";
-                mediaExtractor.setDataSource(mContext, uri, null);
-                int trackCount = mediaExtractor.getTrackCount();
-                int audioTrackindex = -1;
-                int framerate = -1;
-                for (int i = 0; i < trackCount; i++) {
-                    MediaFormat trackFormat = mediaExtractor.getTrackFormat(i);
-                    String mime = trackFormat.getString(MediaFormat.KEY_MIME);
-                    if (mime.startsWith("audio/")) {
-                        mediaExtractor.selectTrack(i);
-                        mediaMuxer = new MediaMuxer(tempAudioPath, MediaMuxer.OutputFormat.MUXER_OUTPUT_MPEG_4);
-                        audioTrackindex = mediaMuxer.addTrack(trackFormat);
-                        mediaMuxer.start();
-                    }
-                    if (mime.startsWith("video/")) {
-                        framerate = trackFormat.getInteger(MediaFormat.KEY_FRAME_RATE);
-                        break;
-                    }
-                }
-
-                if (mediaMuxer == null) {
-                    mContext.revokeUriPermission(uri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
-                    return null;
-                }
-
-                MediaCodec.BufferInfo bufferInfo = new MediaCodec.BufferInfo();
-                bufferInfo.presentationTimeUs = 0;
-
-                ByteBuffer byteBuffer = ByteBuffer.allocate(500 * 1024);
-
-                int i;
-                while ((i = mediaExtractor.readSampleData(byteBuffer, 0)) > 0) {
-                    bufferInfo.offset = 0;
-                    bufferInfo.size = i;
-                    bufferInfo.flags = mediaExtractor.getSampleFlags();
-                    bufferInfo.presentationTimeUs += 1000 * 1000 / framerate;
-                    mediaMuxer.writeSampleData(audioTrackindex, byteBuffer, bufferInfo);
-                    mediaExtractor.advance();
-                }
-
-                mediaExtractor.release();
-                mediaMuxer.stop();
-                mediaMuxer.release();
-
-                mContext.revokeUriPermission(uri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
-                tempAudioFile = new File(tempAudioPath);
-                uri = FileProvider.getUriForFile(mContext, mContext.getApplicationContext().getPackageName() + ".provider", tempAudioFile);
-                mContext.grantUriPermission(PACKAGE_ANKI, uri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
-                uriString = uri.toString();
-            } catch (IOException e) {
-                mContext.revokeUriPermission(uri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
+            mContext.revokeUriPermission(uri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
+            Uri extractedAudioUri = AudioExtractionUtil.extract(mContext, uri, PATH_TEMP_AUDIO);
+            if (extractedAudioUri == null) {
+                File tempAudioFile = new File(PATH_TEMP_AUDIO);
                 if (tempAudioFile != null && tempAudioFile.exists()) {
                     tempAudioFile.delete();
                 }
                 return null;
+            } else {
+                uri = extractedAudioUri;
+                mContext.grantUriPermission(PACKAGE_ANKI, uri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
+                uriString = uri.toString();
             }
         }
         ContentValues cv = new ContentValues();
@@ -409,12 +352,13 @@ public class AnkiDroidHelper {
         try {
             Uri insertedUri = mResolver.insert(Uri.withAppendedPath(FlashCardsContract.AUTHORITY_URI, "media"), cv);
             File insertedFile = new File(insertedUri.getPath());
-            String filePath =  insertedFile.toString();
+            String filePath = insertedFile.toString();
             return filePath.substring(1); // get rid of the "/" at the beginning
         } catch (Exception e) {
             return null;
         } finally {
             mContext.revokeUriPermission(uri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
+            File tempAudioFile = new File(PATH_TEMP_AUDIO);
             if (tempAudioFile != null && tempAudioFile.exists()) {
                 tempAudioFile.delete();
             }
